pyproject.toml
[project]
name = "aerospace-traceability-engine"
version = "0.1.0"
description = "Aerospace Requirements Traceability Engine with LLM-based analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.4.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]
